pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devlogs"
version = "2.0.0"
description = "Developer-focused logging library for Python with OpenSearch integration."
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "Dan Driscoll", email = "dan@thedandriscoll.org"}]
dependencies = [
  "opensearch-py>=2.4.0",
  "click>=8.1.0",
  "typer>=0.9.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.24.0",
  "python-dotenv>=1.0.0",
  "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.26.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
devlogs = "devlogs.wrapper:main"
devlogs-cli = "devlogs.cli:main"
devlogs-collector = "devlogs.collector.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"devlogs.web" = ["static/*.html", "static/*.css", "static/*.js"]

[tool.pytest.ini_options]
markers = [
  "integration: marks tests as integration (deselect with '-m \"not integration\"')",
]
